services/qmark-imagewm/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-magic==0.4.27
PyMuPDF==1.23.22
Pillow==10.2.0
numpy==1.26.4
scipy==1.13.0